// __generated__/CreateStatMutation.graphql.ts
export type stats_insert_input = {
  assists?: number | null | undefined;
  club?: clubs_obj_rel_insert_input | null | undefined;
  club_id?: any | null | undefined;
  created_at?: any | null | undefined;
  goals?: number | null | undefined;
  id?: any | null | undefined;
  matches?: number | null | undefined;
  player?: players_obj_rel_insert_input | null | undefined;
  player_id?: any | null | undefined;
  season?: string | null | undefined;
  updated_at?: any | null | undefined;
};
// ...
export type CreateStatMutation$variables = {
  object: stats_insert_input;
};
export type CreateStatMutation$data = {
  readonly insert_stats_one: {
    readonly assists: number;
    readonly club_id: any;
    readonly goals: number;
    readonly id: any;
    readonly matches: number;
    readonly player_id: any;
    readonly season: string;
  } | null | undefined;
};
export type CreateStatMutation = {
  response: CreateStatMutation$data;
  variables: CreateStatMutation$variables;
};